refactor(nip19): move codecs to own package
