+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/EntityType.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-enum EntityType {
+public enum EntityType {
     NPUB("npub"),
     NSEC("nsec"),
     NOTE("note"),
