+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/codec/Codec.java
@@ -1,4 +1,4 @@
-package org.tbk.nostr.nip19;
+package org.tbk.nostr.nip19.codec;
 
 public interface Codec<T> {
     boolean supports(String hrp, Class<?> clazz);
