+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/codec/util/TlvType.java
@@ -1,11 +1,11 @@
-package org.tbk.nostr.nip19;
+package org.tbk.nostr.nip19.codec.util;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-enum TlvType {
+public enum TlvType {
     SPECIAL((byte) 0),
     RELAY((byte) 1),
     AUTHOR((byte) 2),
